feat(shake.ts): refactor Task and File classes, add tests
This commit refactors the Task and File classes in shake.ts to improve
their structure and readability. The Status type has been changed to an
interface, and the Runnable interface has been replaced with the TaskLike
interface. The Task and File classes now implement TaskLike.

The run methods in Task and File have been updated to return a state,
which is an array of Status objects. The watch function has been updated
to use the new run method.

A new function, topsort, has been added to sort tasks in topological
order. The run function now uses topsort to determine the order in which
tasks should be run.

Finally, a new test file, test.ts, has been added with tests for the
updated Task class and the new run and topsort functions.