feat: Handle file not found error in File task run method
This commit introduces error handling in the run method of the File task.
Previously, if the target file did not exist, fs.stat would throw an
unhandled error. Now, a NotFoundError is caught and handled by setting
the modification time to the Unix epoch, effectively treating the file
as if it has never been modified. This allows tasks dependent on the
non-existent file to run as expected.