feat(shake.ts): refactor topsort to handle cyclic dependencies
The topsort function in shake.ts has been refactored to handle cyclic
dependencies. A new function, findCycle, has been introduced to detect
cycles in the task dependencies. If a cycle is detected, an error is thrown.
The test suite has also been updated to test for a larger set of vertices.